+++ src/main/java/com/atlauncher/gui/tabs/ToolsTab.java
@@ -27,6 +27,7 @@ import com.atlauncher.gui.components.BlankToolPanel;
 import com.atlauncher.gui.components.LogClearerToolPanel;
 import com.atlauncher.gui.components.NetworkCheckerToolPanel;
 import com.atlauncher.gui.components.ServerCheckerToolPanel;
+import com.atlauncher.gui.components.RelaunchInDebugModePanel;
 
 @SuppressWarnings("serial")
 public class ToolsTab extends JPanel implements Tab {
@@ -42,7 +43,7 @@ public class ToolsTab extends JPanel implements Tab {
         mainPanel.add(new NetworkCheckerToolPanel());
         mainPanel.add(new ServerCheckerToolPanel());
         mainPanel.add(new LogClearerToolPanel());
-        mainPanel.add(new BlankToolPanel());
+        mainPanel.add(new RelaunchInDebugModePanel());
         mainPanel.add(new BlankToolPanel());
         mainPanel.add(new BlankToolPanel());