+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 
 -   Fix Edit Mods button not showing on instances
 -   Fix not being able to install older packs without a logging client
+-   Add in Relaunch in debug mode tool for relaunching in debug mode